Changed to use Cosmos DB .NET SDK v3 (#60)
* Changed to use Cosmos DB .NET SDK v3

* Rename `collection` to `container` according to the latest naming convention

* Drop partition options entirely. Still accept 'collectionId' for compat.

Also revved down the Cosmos SDK version a bit to not force people into upgrading if they have older dependencies. (Probably not for the Cosmos SDK specifically, but requiring a minimum of 1.3 instead of 1.19 for dependencies like Azure.Core is reasonable. It does not prevent folks from using newer versions if they want.)

---------

Co-authored-by: Steve Molloy <smolloy@microsoft.com>